Move validation logic outside Swing panel
